Fabrica Improvements - task 2 (updated tests)
