+++ app/routes/settings.js
@@ -1,0 +1,36 @@
+import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
+
+export default Route.extend({
+  can: service(),
+  currentUser: service(),
+  prefixes: service(),
+
+  model() {
+    if (this.can.can('read index')) {
+      let self = this;
+      return this.store
+        .findRecord('provider', 'admin')
+        .then(function (model) {
+          return model;
+        })
+        .catch(function (reason) {
+          console.debug(reason);
+
+          self.get('flashMessages').warning(reason);
+        });
+    }
+  },
+
+  afterModel() {
+    if (this.can.cannot('read index') && this.currentUser) {
+      this.transitionTo('index');
+    }
+  },
+
+  actions: {
+    queryParamsDidChange() {
+      this.refresh();
+    }
+  }
+});
