+++ cypress/integration/organization_admin/settings.test.ts
@@ -1,10 +1,6 @@
 /// <reference types="cypress" />
 /* eslint-disable no-undef */
 
-function escapeRE(string) {
-  return string.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
-}
-
 describe('ACCEPTANCE: ORGANIZATION_ADMIN | SETTINGS', () => {
   const waitTime = 1000;
   const waitTime2 = 2000;
@@ -19,16 +15,16 @@ describe('ACCEPTANCE: ORGANIZATION_ADMIN | SETTINGS', () => {
     cy.wait(waitTime2);
   });
 
-  it('is logged in to homepage', () => {
-    cy.visit('/providers/datacite');
-    cy.url().should('include', '/providers/datacite').then (() => {
+  it('is logged in to settings page', () => {
+    cy.visit('/providers/datacite/settings');
+    cy.url().should('include', '/providers/datacite/settings').then (() => {
       cy.get('h2.work').contains('DataCite');
       cy.get('a#account_menu_link').should('contain', 'DATACITE');
 
       cy.get('ul.nav-tabs li a').contains(/Info/i)
-        .and('have.attr', 'href').and('include', '/providers/datacite/info');
-      cy.get('ul.nav-tabs li.active a').contains(/Settings/i)
         .and('have.attr', 'href').and('include', '/providers/datacite');
+      cy.get('ul.nav-tabs li.active a').contains(/Settings/i)
+        .and('have.attr', 'href').and('include', '/providers/datacite/settings');
       cy.get('ul.nav-tabs li a').contains(/Contacts/i)
         .and('have.attr', 'href').and('include', '/providers/datacite/contacts');
       cy.get('ul.nav-tabs li a').contains(/Repositories/i)
@@ -93,114 +89,6 @@ describe('ACCEPTANCE: ORGANIZATION_ADMIN | SETTINGS', () => {
         // cy.get('[cy-data="secondary-technical"] a').contains(/John\s*Doe/i).and('have.attr', 'href').and('include', 'mailto:John.Doe7426@example.org');
         cy.get('[cy-data="secondary-technical"] a').contains(/John\s*Doe/i).and('have.attr', 'href').and('contain', 'mailto:John.Doe');
       });
-    });
-  });
-
-  it('/info redirects to homepage', () => {
-    cy.visit('/info');
-    cy.url().should('include', '/providers/datacite').then (() => {
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-    });
-  });
-
-  it('/contacts redirects to homepage', () => {
-    cy.visit('/contacts');
-    cy.url().should('include', '/providers/datacite').then (() => {
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-    });
-  });
-
-  it('/repositories redirects to homepage', () => {
-    cy.visit('/repositories');
-    cy.url().should('include', '/providers/datacite').then (() => {
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-    });
-  });
-
-  it('/prefixes redirects to homepage', () => {
-    cy.visit('/prefixes');
-    cy.url().should('include', '/providers/datacite').then (() => {
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-    });
-  });
-
-  it('/dois redirects to homepage', () => {
-    cy.visit('/dois');
-    cy.url().should('include', '/providers/datacite').then (() => {
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-    });
-  });
-
-  it('has password settings page', () => {
-    cy.visit('/providers/datacite/change');
-    cy.url().should('include', '/providers/datacite/change').then (() => {
-
-      cy.get('h2.work').contains('DataCite');
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-
-      cy.get('h3').contains(/Set Password/i);
-
-      cy.get('form').within(($form) => {
-        cy.get('input#password-input-field').should('be.visible');
-        cy.get('input#confirm-password-input-field').should('be.visible');
-        cy.get('button[type=submit]').should('be.visible');
-        cy.get('button').contains(/Cancel/i).should('be.visible')
-
-        cy.get('button').contains(/Cancel/i).click({force: true});
-      });
-    })
-
-    cy.on("url:changed", (newUrl) => {
-      expect(newUrl).to.contain("/providers/datacite");
-    });
-  });
-
-  // TBD - Could do more testing here. Only cursory testing for
-  // presence of field objects. No testing of form behavior yet.
-  it('has an organization settings page', () => {
-    cy.visit('/providers/datacite/edit');
-    cy.url().should('include', '/providers/datacite/edit').then(() => {
-
-      cy.get('h2.work').contains('DataCite');
-      cy.get('a#account_menu_link').should('contain', 'DATACITE');
-
-      cy.get('h3.edit').contains(/Update\s*Organization/);
-
-      cy.get('form').within(($form) => {
-        cy.get('h3.member-results').contains('Organization Information');
-        cy.get('#member-id').should('be.visible');
-        cy.get('#ror-id').should('be.visible');
-        cy.get('#name').should('be.visible');
-        cy.get('#display-name').should('be.visible');
-        cy.get('#system-email').should('be.visible');
-        cy.get('#group-email').should('be.visible');
-        cy.get('#website').should('be.visible');
-        cy.get('#twitter-handle').should('be.visible');
-        cy.get('#country').should('be.visible');
-        cy.get('#organization-type').should('be.visible');
-        cy.get('#focus-area').should('be.visible');
-        cy.get('#description').should('be.visible');
-
-        cy.get('h3.member-results').contains('Contact Information');
-        cy.get('.alert-info').contains(/Contacts are created.*and then assigned roles here./i )
-        cy.get('#service-contact').should('be.visible');
-        cy.get('#secondary-service-contact').should('be.visible');
-        cy.get('#technical-contact').should('be.visible');
-        cy.get('#secondary-technical-contact').should('be.visible');
-        cy.get('.alert-warning').contains(/The contacts entered may receive notifications/i)
-          .within(() => {
-            cy.get('a[href*="privacy.html"]').should('be.visible');
-          }
-        );
-        cy.get('button#update-provider').should('be.visible');
-        cy.get('button').contains(/Cancel/i).should('be.visible');
-
-        cy.get('button').contains(/Cancel/i).click({force: true});
-      })
-    });
-
-    cy.on("url:changed", (newUrl) => {
-      expect(newUrl).to.contain("/providers/datacite");
     });
   });
 });
